Redirect to home on submission of a new question
Comment: You'll need to do a couple of changes here. In the rubric, it is speficied that upon form submission user should be redirected to the home page, while right now it's taken straight to the question. Also, the form validation could be improved, because now it's possible to enter just empty spaces.

Response: Added the validation by adding trim() to option text checking for both. Also changed redirection to / from earlier /questions/:question_id